Assignment 2/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
importlib-metadata==4.8.1
iniconfig==1.1.1
lxml==4.6.3
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-pptx==0.6.19
toml==0.10.2
typing-extensions==3.10.0.2
Wand==0.6.7
XlsxWriter==3.0.1
zipp==3.5.0
